qualquer coisa

+++ src/main/java/br/com/tutorial/domain/embs/Telefone.java
@@ -24,6 +24,6 @@ public class Telefone {
 	
 	private Integer ddd;
 	
-	private Integer numero;
+	private Long numero;
 	
 }
